requirements.txt
Flask==2.1.2
python-dotenv==0.20.0
openai==0.27.0  # Or google-generativeai for Gemi